exercisescorrekt

+++ Exercise2/Program.cs
@@ -7,7 +7,7 @@ int one = int.Parse(input1);
 string input2 = Console.ReadLine();
 int two = int.Parse(input2);
 
-int sum = one / two;
+float sum = (float)one / two;
 
 
 Console.WriteLine($"{sum}");
